refactor: refactored user courses conclusions architecture
