+++ src/app/app.module.ts
@@ -4,7 +4,7 @@ import { UserModule } from '../modules/user/infra/modules/user.module';
 import { UserScoreModule } from '../modules/user-score/infra/modules/user-score.module';
 import { CourseModule } from '../modules/course/infra/modules/course.module';
 import { UserActivitiesAnsweredModule } from '../modules/user-activities-answered/infra/modules/user-activities-answered.module';
-import { UserCoursesConcludedModule } from '../modules/user-courses-concluded/user-courses-concluded.module';
+import { UserCoursesConcludedModule } from '../modules/user-courses-concluded/infra/modules/user-courses-concluded.module';
 import { ActivityModule } from '../modules/activity/infra/modules/activity.module';
 
 @Module({
