+++ src/modules/user-activities-answered/infra/services/user-activities-answered.service.spec.ts
@@ -3,20 +3,20 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { UserActivitiesAnsweredService } from './user-activities-answered.service';
 import { DatabaseModule } from '../../../../database/database.module';
 import { UserModule } from '../../../user/infra/modules/user.module';
-import { User } from '../../user/entity/user.entity';
-import { UserScore } from '../../user-score/entity/user-score.entity';
+import { User } from '../../../user/infra/db/entity/user.entity';
+import { UserScore } from '../../../user-score/infra/db/entity/user-score.entity';
 import { UserScoreService } from '../../../user-score/infra/services/user-score.service';
-import { UserScoreRepository } from '../../user-score/repositories/user-score.repository';
-import { UserRepository } from '../../user/repositories/user.repository';
-import { JWTProvider } from '../../user/providers/jwt.provider';
+import { UserScoreRepository } from '../../../user-score/infra/db/repositories/user-score.repository';
+import { UserRepository } from '../../../user/infra/db/repositories/user.repository';
+import { JWTProvider } from '../../../user/infra/providers/jwt.provider';
 import { UserNotFoundException } from '../../../user/domain/dtos/errors/UserNotFound.exception';
 import { ActivityNotFoundException } from '../../domain/dtos/errors/ActivityNotFound.exception';
 import { ActivityAlreadyAnsweredException } from '../../domain/dtos/errors/ActivityAlreadyAnswered.exception';
 import { WrongAnswerException } from '../../domain/dtos/errors/WrongAnswer.exception';
-import { UserActivityAnsweredRepository } from '../repositories/user-activities-answered.repository';
-import { ActivityRepository } from '../../activity/repositories/activity.repository';
-import { UserCourseConcludedRepository } from '../../../user-courses-concluded/repository/user-courses-concluded.repository';
-import { CourseRepository } from '../../course/repositories/course.repository';
+import { UserActivityAnsweredRepository } from '../db/repositories/user-activities-answered.repository';
+import { ActivityRepository } from '../../../activity/infra/db/repositories/activity.repository';
+import { UserCourseConcludedRepository } from '../../../user-courses-concluded/infra/db/repositories/user-courses-concluded.repository';
+import { CourseRepository } from '../../../course/infra/db/repositories/course.repository';
 import { TestHelper } from '../../../../../test/helpers/dbInstanceHelper';
 
 describe('UserActivitiesAnsweredService', () => {
