+++ src/modules/user-courses-concluded/domain/dtos/repositories/user-courses-concluded.repository.ts
@@ -1,0 +1,9 @@
+import { UserCourseConcluded } from "../../../infra/db/entity/user-courses-concluded.entity";
+
+export interface UserCoursesConcludedRepositoryInterface {
+    findConcludedCourse(user_id: number, course_id: number): Promise<UserCourseConcluded>;
+
+    countUserConcludedCourses(user_id: number): Promise<number>;
+
+    countCourseGlobalConclusions(course_id: number): Promise<number>;
+}
