+++ src/modules/user-courses-concluded/infra/db/entity/user-courses-concluded.entity.spec.ts
