+++ src/modules/user-courses-concluded/infra/db/repositories/user-courses-concluded.repository.spec.ts
