+++ src/modules/user/infra/modules/user.module.ts
@@ -18,7 +18,7 @@ import { UserScore } from '../../../user-score/entity/user-score.entity';
 import { UserRepository } from '../db/repositories/user.repository';
 import { UserScoreRepository } from '../../../user-score/repositories/user-score.repository';
 import { AuthenticationMiddleware } from '../http/middlewares/auth.middleware';
-import { UserCourseConcludedRepository } from '../../../user-courses-concluded/repository/user-courses-concluded.repository';
+import { UserCourseConcludedRepository } from '../../../user-courses-concluded/repositories/user-courses-concluded.repository';
 
 @Module({
   imports: [
